mergerd LMA-1 to feature/book and reogranization of project structire
