--- src/main/java/pl/sda/librarymanagementapp/Book/BookCreateService.java
@@ -1,4 +1,0 @@
-package pl.sda.librarymanagementapp.Book;
-
-public class BookCreateService {
-}
