--- src/main/java/pl/sda/librarymanagementapp/Book/BookDto.java
@@ -1,21 +1,0 @@
-package pl.sda.librarymanagementapp.Book;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class BookDto {
-
-    private Long id;
-    private String title;
-    private String author;
-    private String publisher;
-    private String publicationYear;
-    private String placeOfPublication;
-    private String isbn;
-}
