--- src/main/java/pl/sda/librarymanagementapp/Book/BookMapper.java
@@ -1,12 +1,0 @@
-package pl.sda.librarymanagementapp.Book;
-
-import org.mapstruct.Mapper;
-
-@Mapper
-public interface BookMapper {
-
-    Book toBook(BookDto bookDto);
-    BookDto toBookDto(Book book);
-
-
-}
