+++ src/main/java/pl/sda/librarymanagementapp/Config.java
@@ -5,9 +5,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestTemplate;
-import pl.sda.librarymanagementapp.Book.Book;
-import pl.sda.librarymanagementapp.Book.BookDto;
-import pl.sda.librarymanagementapp.Book.BookMapper;
+import pl.sda.librarymanagementapp.domain.Book;
+import pl.sda.librarymanagementapp.model.book.BookDto;
+import pl.sda.librarymanagementapp.model.mapper.BookMapper;
 
 @Configuration
 public class Config {
