+++ src/main/java/pl/sda/librarymanagementapp/domain/Book.java
@@ -3,15 +3,20 @@ package pl.sda.librarymanagementapp.domain;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import lombok.RequiredArgsConstructor;
+import pl.sda.librarymanagementapp.domain.LibraryUser;
 
-import javax.persistence.*;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.ManyToOne;
 
 @Entity(name = "books")
+@AllArgsConstructor
 @Data
 @NoArgsConstructor
-@AllArgsConstructor
-
 public class Book {
+
     @Id
     @GeneratedValue
     private Long id;
@@ -19,10 +24,11 @@ public class Book {
     private String title;
     private String author;
     private String publisher;
-    private String isbn;
     private String publicationYear;
-    private String placeofPublication;
+    private String placeOfPublication;
+    private String isbn;
 
     @ManyToOne
     private LibraryUser libraryUser;
+
 }
