+++ src/main/java/pl/sda/librarymanagementapp/model/book/BookController.java
@@ -1,10 +1,12 @@
-package pl.sda.librarymanagementapp.Book;
+package pl.sda.librarymanagementapp.model.book;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
+import pl.sda.librarymanagementapp.domain.Book;
+import pl.sda.librarymanagementapp.model.mapper.BookMapper;
 
 @RestController
 @RequiredArgsConstructor
