+++ src/main/java/pl/sda/librarymanagementapp/model/book/BookCreateService.java
@@ -1,0 +1,4 @@
+package pl.sda.librarymanagementapp.model.book;
+
+public class BookCreateService {
+}
