+++ src/main/java/pl/sda/librarymanagementapp/model/book/BookSourceResponse.java
@@ -1,4 +1,4 @@
-package pl.sda.librarymanagementapp.Book;
+package pl.sda.librarymanagementapp.model.book;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
